Add a run-harness that (attempts to) gracefully stop a program
This script behaves like 'timeout' command however instead of
sending a signal it gracefully exits the running command using an
exit(0) called from within the process.

The reason for this is because compiler-instrumented binaries
(usually) only dump their records on an exit() call.  Programs
that don't elegantly route signals ending with an exit() call
will have their instrumented results discarded and not written
to disk.

Therefore, in the case of running DOSBox on a headless system,
the only way capture instrumented output is to strap a debugger
to the processes and call exit(0) from within.